Rename hook with tx prefix
