+++ common/v2/utils/index.ts
@@ -52,4 +52,5 @@ export { toArray } from './toArray';
 export { objToString } from './objToString';
 export * from './constants';
 export { bigify } from './bigify';
-export { useTxSend } from './useTxSend/useTxSend';
+export { useTxSend } from './useTxSend';
+export { useTxMulti } from './useTxMulti';
