--- common/v2/utils/useMultiTx/index.ts
@@ -1,1 +1,0 @@
-export { useMultiTx, TxParcel } from './useMultiTx';
