+++ common/v2/utils/useTxMulti/index.ts
@@ -1,0 +1,1 @@
+export { useTxMulti, TxParcel } from './useTxMulti';
